SUMMARY:  Merging back changes to SuiteSparse to address compiler warnings from Windows
     BUGZILLA BUG #(s):  None.

     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  N/A

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  N

     3)  Did you modify or add any source or build system files?  N
     3a) Did you rebuild the libraries and executables?  Y
     3b) Did you rerun the test suite on the packages that you changed and
         all packages that depend on the modified packages?  Y
     3c) Did you change the required or optional dependencies for any
         package?  N


     DESCRIPTION OF CHANGES:

     Merging back changes from release branch.

(checklist version 2009-07-14)

+++ packages/amesos/src/SuiteSparse/AMD/Source/amesos_amd_2.c
@@ -600,7 +600,7 @@ GLOBAL void AMD_2
     }
     else
     {
-	dense = alpha * sqrt ((double) n) ;
+	dense = (int) ( alpha * sqrt ((double) n) ) ;
     }
     dense = MAX (16, dense) ;
     dense = MIN (n,  dense) ;
